+++ feature/onboard/src/main/java/kr/co/onboard/OnBoardViewModel.kt
@@ -1,4 +1,4 @@
-package kr.co.onboard.ui
+package kr.co.onboard
 
 import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
